Module aws_sdk_connectcontactlens::model

Modules

categories

See Categories

category_details

See CategoryDetails

character_offsets

See CharacterOffsets

issue_detected

See IssueDetected

point_of_interest

See PointOfInterest

realtime_contact_analysis_segment

See RealtimeContactAnalysisSegment

transcript

See Transcript

Structs

Categories

Provides the category rules that are used to automatically categorize contacts based on

CategoryDetails

Provides information about the category rule that was matched.

CharacterOffsets

For characters that were detected as issues, where they occur in the transcript.

IssueDetected

Potential issues that are detected based on an artificial intelligence analysis of each

PointOfInterest

The section of the contact audio where that category rule was detected.

RealtimeContactAnalysisSegment

An analyzed segment for a real-time analysis session.

Transcript

A list of messages in the session.

Enums

SentimentValue